refactored sign to have parties verify outputted signatures
